+++ Source/PlaneMode/Log.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace PlaneMode
+﻿namespace PlaneMode
 {
     internal static class Log
     {
@@ -11,49 +9,59 @@ namespace PlaneMode
             Level = LogLevel.Info;
         }
 
-        public static void Error(string format, params object[] args)
+        public static void Error(string message)
         {
+            Level = Config.Instance.LogLevel;
+
             if ((byte)LogLevel.Error <= (byte)Level)
             {
-                UnityEngine.Debug.LogError(Prefix(LogLevel.Error) + String.Format(format, args));
+                UnityEngine.Debug.LogError($"{Prefix(LogLevel.Error)} {message}");
             }
         }
 
-        public static void Warning(string format, params object[] args)
+        public static void Warning(string message)
         {
+            Level = Config.Instance.LogLevel;
+
             if ((byte)LogLevel.Warning <= (byte)Level)
             {
-                UnityEngine.Debug.LogWarning(Prefix(LogLevel.Warning) + String.Format(format, args));
+                UnityEngine.Debug.LogWarning($"{Prefix(LogLevel.Warning)} {message}");
             }
         }
 
-        public static void Info(string format, params object[] args)
+        public static void Info(string message)
         {
+            Level = Config.Instance.LogLevel;
+
             if ((byte)LogLevel.Info <= (byte)Level)
             {
-                UnityEngine.Debug.Log(Prefix(LogLevel.Info) + String.Format(format, args));
+                UnityEngine.Debug.Log($"{Prefix(LogLevel.Info)} {message}");
             }
         }
 
-        public static void Debug(string format, params object[] args)
+        public static void Debug(string message)
         {
+            Level = Config.Instance.LogLevel;
+
             if ((byte)LogLevel.Debug <= (byte)Level)
             {
-                UnityEngine.Debug.Log(Prefix(LogLevel.Debug) + String.Format(format, args));
+                UnityEngine.Debug.Log($"{Prefix(LogLevel.Debug)} {message}");
             }
         }
 
-        public static void Trace(string format, params object[] args)
+        public static void Trace(string message)
         {
+            Level = Config.Instance.LogLevel;
+
             if ((byte)LogLevel.Trace <= (byte)Level)
             {
-                UnityEngine.Debug.Log(Prefix(LogLevel.Trace) + String.Format(format, args));
+                UnityEngine.Debug.Log($"{Prefix(LogLevel.Trace)} {message}");
             }
         }
 
         private static string Prefix(LogLevel level)
         {
-            return String.Format("[PlaneMode] [{0}]: ", level);
+            return $"[PlaneMode] [{level}]:";
         }
     }
 }